Fix error when bashrc or profile don't exist

+++ setup/cleanup.sh
@@ -3,10 +3,10 @@
 # check if files are not symbolic, meaning they
 # are the default files and not symlinked to the
 # dotfiles repo files
-if [[ ! -h ~/.profile ]]; then
+if [[ -f ~/.profile ]] && [[ ! -h ~/.profile ]]; then
   mv ~/.profile ~/.profile.old 
 fi
 
-if [[ ! -h ~/.bashrc ]]; then
+if [[ -f ~/.bashrc ]] && [[ ! -h ~/.bashrc ]]; then
   mv ~/.bashrc ~/.bashrc.old 
 fi
